+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-## [0.5.0] 2023-08-10
+### Changed
+- Updated to intellij platform `2025.1`
+
 ## [0.6.0] 2024-12-22
 
 ### Changed